tutoriales_01-06/tutorial06_flask01/requirements.txt
asgiref==3.11.1
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
